Re-layout PFSAgent to support eventual callable package variant
The program that was pfsagentd is now a sub-directory of new
package pfsagent. Currently, the only callable API in package
pfsagent is Daemon() of which the pfsagentd "main" program
launches.
