+++ pfsagent/functional_test.go
@@ -1,4 +1,4 @@
-package main
+package pfsagent
 
 import (
 	"testing"
