+++ pfsagent/globals.go
@@ -1,4 +1,4 @@
-package main
+package pfsagent
 
 import (
 	"container/list"
