+++ pfsagent/pfsagentd/Makefile
@@ -1,0 +1,3 @@
+gosubdir := github.com/swiftstack/ProxyFS/pfsagent/pfsagentd
+
+include ../../GoMakefile
