+++ pfsagent/pfsagentd/dummy_test.go
@@ -1,0 +1,8 @@
+package main
+
+import (
+	"testing"
+)
+
+func TestDummy(t *testing.T) {
+}
