+++ pfsagent/swift_proxy_emulator_test.go
@@ -1,4 +1,4 @@
-package main
+package pfsagent
 
 import (
 	"context"
