--- pfsagentd/Makefile
@@ -1,3 +1,0 @@
-gosubdir := github.com/swiftstack/ProxyFS/pfsagentd
-
-include ../GoMakefile
